--- a/__tests__/main.runTaskListOnce.test.ts
+++ b/__tests__/main.runTaskListOnce.test.ts
@@ -12,7 +12,7 @@ test("runs jobs", () =>
     // Schedule a job
     const start = new Date();
     await pgClient.query(
-      `select * from graphile_worker.add_job('job1', '{"a": 1}')`
+      `select * from graphile_worker.add_job('job1', '{"a": 1}', queue_name := 'myqueue')`
     );
 
     // Assert that it has an entry in jobs / job_queues
@@ -61,7 +61,7 @@ test("schedules errors for retry", () =>
     // Schedule a job
     const start = new Date();
     await pgClient.query(
-      `select * from graphile_worker.add_job('job1', '{"a": 1}')`
+      `select * from graphile_worker.add_job('job1', '{"a": 1}', queue_name := 'myqueue')`
     );
 
     {
@@ -129,7 +129,7 @@ test("retries job", () =>
 
     // Add the job
     await pgClient.query(
-      `select * from graphile_worker.add_job('job1', '{"a": 1}')`
+      `select * from graphile_worker.add_job('job1', '{"a": 1}', queue_name := 'myqueue')`
     );
     let counter = 0;
     const job1: Task = jest.fn(() => {
@@ -513,7 +513,7 @@ test("job details are reset if not specified in update", () =>
       max_attempts: 25,
       payload: {},
       task_identifier: "job1",
-      queue_name: "queue1",
+      queue_name: null,
     });
 
     // update job with new details
@@ -635,7 +635,7 @@ test("runs jobs asynchronously", () =>
     // Schedule a job
     const start = new Date();
     await pgClient.query(
-      `select * from graphile_worker.add_job('job1', '{"a": 1}')`
+      `select * from graphile_worker.add_job('job1', '{"a": 1}', queue_name := 'myqueue')`
     );
 
     // Run the task
@@ -703,7 +703,7 @@ test("runs jobs in parallel", () =>
     // Schedule 5 jobs
     const start = new Date();
     await pgClient.query(
-      `select graphile_worker.add_job('job1', '{"a": 1}') from generate_series(1, 5)`
+      `select graphile_worker.add_job('job1', '{"a": 1}', queue_name := 'queue_' || s::text) from generate_series(1, 5) s`
     );
 
     // Run the task