Before:

```
Timing startup/shutdown time...
... it took 67ms

Scheduling 20000 jobs


Timing 20000 job execution...
Found 999!

... it took 3210ms
Jobs per second: 6363.16


Testing latency...
[core] INFO: Worker connected and looking for jobs... (task names: 'latency')
Beginning latency test
Latencies - min: 2.47ms, max: 12.00ms, avg: 2.72ms
Done
Done in 10.52s.
```

After:

```
Timing startup/shutdown time...
... it took 65ms

Scheduling 20000 jobs


Timing 20000 job execution...
Found 999!

... it took 2033ms
Jobs per second: 10165.04


Testing latency...
[core] INFO: Worker connected and looking for jobs... (task names: 'latency')
Beginning latency test
Latencies - min: 2.42ms, max: 12.37ms, avg: 2.71ms
Done
Done in 9.03s.
```