@benjie 2020-02-04
@gregplaysguitar Your suspicion about only using `job_queues` when necessary has panned out: we now max out at 10kjps versus 6.3kjps previously.